HZN-1510: Remove DelegatingVertexEdgeProvider
